Trying out a new concept: autodetection
This way you only need to configure `elmFolder` and `outputFile`. All main
files will be automatically detected and built into one outputFile,
removing the need to configure anything and the duplication of the elm
runtime that happens if you have multiple files.

Also, this works for files you add after you start "brunch watch", and
it could also auto-detect removal of files.

The only issue so far is that it will build once per discovered main
file at boot since there is no way to know when brunch has called the
plugin with all possible file names on boot.

Haven't written any tests for this since it's just a prototype
implementation to try the idea.